refactor mapsService
