+++ codice/.sonarlint/storage/h2/sq-ide.lock.db
@@ -1,6 +1,6 @@
 #FileLock
-#Thu Apr 16 07:56:12 CEST 2026
+#Fri Apr 17 07:45:46 CEST 2026
 hostName=localhost
-id=19d94dc7f9431ee172ced5068542fea386133be39fa
+id=19d99f951c8d188ee5e99eef9e257d2ba23d2cdbc75
 method=file
-server=localhost\:55169
+server=localhost\:50088
