+++ codice/PMG-backend/src/main/java/pmg/backend/prenotazione/StatoPrenotazione.java
@@ -2,7 +2,7 @@ package pmg.backend.prenotazione;
 
 public enum StatoPrenotazione {
     attiva,         // Prenotata (attesa entro 10 min)
-    inCorso,       // Utente entrato (timer avviato)
+    inCorso,        // Utente entrato (timer avviato)
     parcheggiato,   // Utente ha raggiunto e confermato il posto
     pagato,         // Saldo effettuato (pronto per uscire, 10 min max)
     conclusa,       // Utente uscito (posto liberato)
